+++ src/github/clone_repo.rs
@@ -1,0 +1,3 @@
+use super::GithubRepo;
+
+pub fn clone_repo(repo: &GithubRepo, i: usize, len: usize) {}
